[cleanup] Remove legacy author usage in javadocs, read git history instead
This type of author is problematic in general as it leads to blaming author listed for code quality when code could have changed by many others and not be the authors fault.  In OSS and in general, its better to let blame land in git history not against the code directly in this way to avoid such author blaming incorrectly and to be more friendly in that developers or even users look at the code and focus on code issues not author relationship.