[cleanup] Remove @since tags as many have no direct relationship to code
The since tag is ok mostly but the release notes should show when items were made available and someone looking at this should be already using this version so its mention is pointless.  Further most of the ones here were not tied to any version but rather timestamps.  Given we are full version 2 and for this type of project would not use this for any further coding efforts and it was not consistently followed, just removing them.